pyproject.toml
[project]
name = "tsc-benchmark"
version = "0.1.0"
description = "Neural network benchmarking framework for time series classification"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "torchaudio>=2.9.0",
    "aeon>=1.3.0",
    "numpy>=1.24.0,<2.3.0",
    "pandas>=2.3.0",
    "scikit-learn>=1.5.0",
    "pyyaml>=6.0.2",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "tqdm>=4.67.0",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "black>=25.12.0",
    "ruff>=0.14.0",
    "pytest>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
target-version = ["py313"]

[tool.ruff]
line-length = 120
target-version = "py313"
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
